+++ src/webserver_main.h
@@ -5,5 +5,6 @@ void Param_setup();
 void Webserver_HandleWifiSetup();
 void Webserver_HandleWifiStatus();
 void Webserver_CheckWifiSetupFallback();
+void Webserver_ClearPendingWifiCredentials();
 void Webserver_HandleWifiScan();
 void Webserver_HandleWifiScanResults();